third_party/blink/web_tests/external/wpt/web-locks/bfcache/held.tentative.https.html

<!DOCTYPE html>
<meta charset=utf-8>
<meta name="timeout" content="long">
<title>Web Locks API: bfcache</title>
<link rel=help href="https://wicg.github.io/web-locks/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="/common/dispatcher/dispatcher.js"></script>
<script src="/html/browsers/browsing-the-web/back-forward-cache/resources/helper.sub.js"></script>
<script type="module">
  import { runWebLocksBfcacheTest } from "./helpers.js";

  runWebLocksBfcacheTest({
    funcBeforeNavigation: () => {
      navigator.locks.request(uniqueNameByQuery(), () => new Promise(() => { }));
    },
    shouldBeCached: false
  }, "A held lock on the main thread must prevent bfcache");

  runWebLocksBfcacheTest({
    funcBeforeNavigation: async () => {
      window.worker = new Worker("/web-locks/resources/worker.js");
      await postToWorkerAndWait(worker, { op: "request", name: uniqueNameByQuery() });
    },
    shouldBeCached: false
  }, "A held lock on a worker must prevent bfcache");

  runWebLocksBfcacheTest({
    funcBeforeNavigation: async () => {
      window.worker = new Worker("/web-locks/resources/parentworker.js");
      await postToWorkerAndWait(worker, { op: "request", name: uniqueNameByQuery() });
    },
    shouldBeCached: false
  }, "A held lock on a nested worker must prevent bfcache");
</script>
